+++ spec/services/insights/generate/repository_service_spec.rb
@@ -4,14 +4,16 @@ describe Insights::Generate::RepositoryService, type: :service do
   subject(:service_call) { described_class.call(insightable: insightable) }
 
   let!(:repository) { create :repository }
-  let!(:pr2) { create :pull_request, repository: repository, entity: entity1 }
-  let!(:pr3) { create :pull_request, repository: repository, pull_merged_at: 10.seconds.after, entity: entity2 }
+  let!(:pr2) { create :pull_request, repository: repository, entity: entity1, changed_loc: 13 }
+  let!(:pr3) {
+    create :pull_request, repository: repository, pull_merged_at: 10.seconds.after, entity: entity2, changed_loc: 11
+  }
   let!(:entity1) { create :entity, external_id: '1' }
   let!(:entity2) { create :entity, external_id: '2' }
   let(:insightable) { repository }
 
   before do
-    create :pull_request, repository: repository, entity: entity1
+    create :pull_request, repository: repository, entity: entity1, changed_loc: 12
     create :pull_requests_comment, entity: entity1, pull_request: pr3
     create :pull_requests_comment, entity: entity2, pull_request: pr2
 
@@ -140,7 +142,10 @@ describe Insights::Generate::RepositoryService, type: :service do
               average_open_pr_comments: true,
               average_review_seconds: true,
               average_merge_seconds: true,
-              review_involving: true
+              review_involving: true,
+              changed_loc: true,
+              average_changed_loc: true,
+              average_reviewed_loc: true
             }
             repository.company.save!
           end
@@ -158,6 +163,9 @@ describe Insights::Generate::RepositoryService, type: :service do
             expect(last_insight.average_merge_seconds).to eq 10
             expect(last_insight.average_open_pr_comments).to eq 3
             expect(last_insight.review_involving).to eq 100
+            expect(last_insight.changed_loc).to eq 11
+            expect(last_insight.average_changed_loc).to eq 11
+            expect(last_insight.average_reviewed_loc).to eq 13
           end
         end
       end